put getters and setters, chance type

+++ src/main/java/com/project/cinema_manager/repository/FilmRepository.java
@@ -3,6 +3,6 @@ package com.project.cinema_manager.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.project.cinema_manager.model.Film;
 
-public interface FilmRepository extends JpaRepository<Film, Long> {
+public interface FilmRepository extends JpaRepository<Film, Integer> {
 
 }
